@@ -5,8 +5,8 @@ interface CardProps {
   children: ReactElement[] | ReactElement;
 }
 
-export const Card: FC<CardProps> = props => (
-  <div className="card" style={{ maxWidth: '650px' }}>
+export const Card: FC<CardProps> = (props) => (
+  <div className="card">
     <div className="card-header bg-primary" style={{ color: '#fff' }}>
       <h5>{props.title}</h5>
     </div>
