@@ -5,7 +5,7 @@ interface FormBtnProps {
   children: string;
 }
 
-export const FormBtn: FC<FormBtnProps> = props => (
+export const FormBtn: FC<FormBtnProps> = (props) => (
   <button onClick={props.onClick} style={{ float: 'right', marginBottom: 10 }} className="btn btn-success">
     {props.children}
   </button>
