@@ -6,7 +6,7 @@ import { MongoError } from 'mongodb';
 const mongoose = require('mongoose');
 mongoose.Promise = global.Promise;
 let MONGO_URL: string;
-const MONGO_LOCAL_URL = 'mongodb://localhost/reactreadinglist-auth';
+const MONGO_LOCAL_URL = 'mongodb://localhost/yahtzee';
 
 if (process.env.MONGODB_URI) {
   mongoose.connect(process.env.MONGODB_URI, { useNewUrlParser: true });
